+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Breaking changes:
 
 Bux fixes:
 * [OLMIS-7496](https://openlmis.atlassian.net/browse/OLMIS-7496): Fix CREATE_ORDER right not working  for users who had Approve requisition
+* [OLMIS-7534](https://openlmis.atlassian.net/browse/OLMIS-7534): OLMIS-7534 Fix unique code validation for existing orderable
 
 
 15.2.2 / 2021-10-29